Added a new cmdlet to unregister application servers.

+++ src/Cohesity.Powershell/Cmdlets/Alert/GetCohesityAlert.cs
@@ -47,7 +47,7 @@ namespace Cohesity.Powershell.Cmdlets.Cluster
         /// </para> 
         /// </summary>
         [Parameter(Mandatory = true)]
-        public int? MaxAlerts { get; set; }
+        public long? MaxAlerts { get; set; }
 
         /// <summary>
         /// <para type="description">
@@ -79,7 +79,7 @@ namespace Cohesity.Powershell.Cmdlets.Cluster
         /// </para>
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int[] ResolutionIds { get; set; }
+        public long[] ResolutionIds { get; set; }
 
         /// <summary>
         /// <para type="description">
