+++ src/Cohesity.Powershell/Cmdlets/Cluster/GetCohesityClusterPartition.cs
@@ -46,7 +46,7 @@ namespace Cohesity.Powershell.Cmdlets.Cluster
         /// </para>
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int[] Ids { get; set; }
+        public long[] Ids { get; set; }
 
         /// <summary>
         /// <para type="description">
