+++ src/Cohesity.Powershell/Cmdlets/Recovery/FindCohesityObjectsForRestore.cs
@@ -70,7 +70,7 @@ namespace Cohesity.Powershell.Cmdlets.Recovery
         /// </para>
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int? StartTime { get; set; }
+        public long? StartTime { get; set; }
 
         /// <summary>
         /// <para type="description">
@@ -80,7 +80,7 @@ namespace Cohesity.Powershell.Cmdlets.Recovery
         /// </para>
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int? EndTime { get; set; }
+        public long? EndTime { get; set; }
 
         /// <summary>
         /// <para type="description">
@@ -88,7 +88,7 @@ namespace Cohesity.Powershell.Cmdlets.Recovery
         /// </para> 
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int[] JobIds { get; set; }
+        public long[] JobIds { get; set; }
 
         /// <summary>
         /// <para type="description">
@@ -96,7 +96,7 @@ namespace Cohesity.Powershell.Cmdlets.Recovery
         /// </para> 
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int[] RegisteredSourceIds { get; set; }
+        public long[] RegisteredSourceIds { get; set; }
 
         /// <summary>
         /// <para type="description">
@@ -104,7 +104,7 @@ namespace Cohesity.Powershell.Cmdlets.Recovery
         /// </para> 
         /// </summary>
         [Parameter(Mandatory = false)]
-        public int[] StorageDomainIds { get; set; }
+        public long[] StorageDomainIds { get; set; }
 
         protected override void BeginProcessing()
         {
